+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Types of changes:
 
 ## [Unreleased]
 
+### Added
+
+- Setting to change the cursor type appearing in insert mode. The default is a "bar" cursor, like in Obsidian with Vim bindings.
+
 ### Changed
 
 - Update codemirror-helix to 0.0.9